clean up code based on liniting
